Adds FinancialConnections with listener implementation.
