+++ financial-connections-example/build.gradle
@@ -51,6 +51,7 @@ dependencies {
     implementation libs.androidx.appCompat
     implementation libs.androidx.coreKtx
     implementation libs.androidx.lifecycle
+    implementation libs.androidx.liveDataKtx
     implementation libs.compose.activity
     implementation libs.compose.material
     implementation libs.compose.liveData
