+++ financial-connections/api/financial-connections.api
@@ -12,9 +12,20 @@ public final class com/stripe/android/financialconnections/ComposableSingletons$
 	public final fun getLambda-1$financial_connections_release ()Lkotlin/jvm/functions/Function2;
 }
 
+public final class com/stripe/android/financialconnections/FinancialConnections {
+	public static final field $stable I
+	public static final field INSTANCE Lcom/stripe/android/financialconnections/FinancialConnections;
+	public static final fun clearEventListener ()V
+	public static final fun setEventListener (Lkotlin/jvm/functions/Function1;)V
+}
+
 public final class com/stripe/android/financialconnections/FinancialConnectionsSheet {
 	public static final field $stable I
 	public static final field Companion Lcom/stripe/android/financialconnections/FinancialConnectionsSheet$Companion;
+	public static final fun create (Landroidx/activity/ComponentActivity;Lcom/stripe/android/financialconnections/FinancialConnectionsSheetResultCallback;)Lcom/stripe/android/financialconnections/FinancialConnectionsSheet;
+	public static final fun create (Landroidx/fragment/app/Fragment;Lcom/stripe/android/financialconnections/FinancialConnectionsSheetResultCallback;)Lcom/stripe/android/financialconnections/FinancialConnectionsSheet;
+	public static final fun createForBankAccountToken (Landroidx/activity/ComponentActivity;Lcom/stripe/android/financialconnections/FinancialConnectionsSheetResultForTokenCallback;)Lcom/stripe/android/financialconnections/FinancialConnectionsSheet;
+	public static final fun createForBankAccountToken (Landroidx/fragment/app/Fragment;Lcom/stripe/android/financialconnections/FinancialConnectionsSheetResultForTokenCallback;)Lcom/stripe/android/financialconnections/FinancialConnectionsSheet;
 	public final fun present (Lcom/stripe/android/financialconnections/FinancialConnectionsSheet$Configuration;)V
 }
 
@@ -203,6 +214,35 @@ public abstract interface class com/stripe/android/financialconnections/Financia
 
 public abstract interface class com/stripe/android/financialconnections/FinancialConnectionsSheetResultForTokenCallback {
 	public abstract fun onFinancialConnectionsSheetResult (Lcom/stripe/android/financialconnections/FinancialConnectionsSheetForTokenResult;)V
+}
+
+public final class com/stripe/android/financialconnections/analytics/FinancialConnectionsEvent {
+	public static final field $stable I
+	public final fun getMetadata ()Lcom/stripe/android/financialconnections/analytics/FinancialConnectionsEvent$Metadata;
+	public final fun getName ()Lcom/stripe/android/financialconnections/analytics/FinancialConnectionsEvent$Name;
+}
+
+public final class com/stripe/android/financialconnections/analytics/FinancialConnectionsEvent$ErrorCode : java/lang/Enum {
+	public static final field INSTITUTION_UNAVAILABLE_PLANNED Lcom/stripe/android/financialconnections/analytics/FinancialConnectionsEvent$ErrorCode;
+	public final fun getValue ()Ljava/lang/String;
+	public static fun valueOf (Ljava/lang/String;)Lcom/stripe/android/financialconnections/analytics/FinancialConnectionsEvent$ErrorCode;
+	public static fun values ()[Lcom/stripe/android/financialconnections/analytics/FinancialConnectionsEvent$ErrorCode;
+}
+
+public final class com/stripe/android/financialconnections/analytics/FinancialConnectionsEvent$Metadata {
+	public static final field $stable I
+	public fun <init> ()V
+	public final fun getErrorCode ()Lcom/stripe/android/financialconnections/analytics/FinancialConnectionsEvent$ErrorCode;
+	public final fun getInstitutionName ()Ljava/lang/String;
+	public final fun getManualEntry ()Ljava/lang/Boolean;
+	public final fun toMap ()Ljava/util/Map;
+}
+
+public final class com/stripe/android/financialconnections/analytics/FinancialConnectionsEvent$Name : java/lang/Enum {
+	public static final field CONSENT_ACQUIRED Lcom/stripe/android/financialconnections/analytics/FinancialConnectionsEvent$Name;
+	public final fun getValue ()Ljava/lang/String;
+	public static fun valueOf (Ljava/lang/String;)Lcom/stripe/android/financialconnections/analytics/FinancialConnectionsEvent$Name;
+	public static fun values ()[Lcom/stripe/android/financialconnections/analytics/FinancialConnectionsEvent$Name;
 }
 
 public final class com/stripe/android/financialconnections/exception/AppInitializationError : com/stripe/android/core/exception/StripeException {
