+++ financial-connections/src/main/java/com/stripe/android/financialconnections/FinancialConnectionsSheetViewModel.kt
@@ -14,8 +14,8 @@ import com.stripe.android.financialconnections.FinancialConnectionsSheetViewEffe
 import com.stripe.android.financialconnections.FinancialConnectionsSheetViewEffect.OpenAuthFlowWithUrl
 import com.stripe.android.financialconnections.FinancialConnectionsSheetViewEffect.OpenNativeAuthFlow
 import com.stripe.android.financialconnections.analytics.FinancialConnectionsAnalyticsTracker
-import com.stripe.android.financialconnections.analytics.FinancialConnectionsEvent.Error
 import com.stripe.android.financialconnections.analytics.FinancialConnectionsEventReporter
+import com.stripe.android.financialconnections.analytics.logError
 import com.stripe.android.financialconnections.browser.BrowserManager
 import com.stripe.android.financialconnections.di.APPLICATION_ID
 import com.stripe.android.financialconnections.di.DaggerFinancialConnectionsSheetComponent
@@ -113,21 +113,22 @@ internal class FinancialConnectionsSheetViewModel @Inject constructor(
                 )
             }
         } else {
-            val authFlowStatus = if (nativeAuthFlowEnabled) {
-                AuthFlowStatus.NONE
+            if (nativeAuthFlowEnabled) {
+                setState {
+                    copy(
+                        manifest = sync.manifest,
+                        webAuthFlowStatus = AuthFlowStatus.NONE,
+                        viewEffect = OpenNativeAuthFlow(initialArgs.configuration, sync)
+                    )
+                }
             } else {
-                AuthFlowStatus.ON_EXTERNAL_ACTIVITY
-            }
-            setState {
-                copy(
-                    manifest = sync.manifest,
-                    webAuthFlowStatus = authFlowStatus,
-                    viewEffect = if (nativeAuthFlowEnabled) {
-                        OpenNativeAuthFlow(initialArgs.configuration, sync)
-                    } else {
-                        OpenAuthFlowWithUrl(sync.manifest.hostedAuthUrl)
-                    }
-                )
+                setState {
+                    copy(
+                        manifest = sync.manifest,
+                        webAuthFlowStatus = AuthFlowStatus.ON_EXTERNAL_ACTIVITY,
+                        viewEffect = OpenAuthFlowWithUrl(sync.manifest.hostedAuthUrl)
+                    )
+                }
             }
         }
     }
@@ -135,7 +136,12 @@ internal class FinancialConnectionsSheetViewModel @Inject constructor(
     private fun logNoBrowserAvailableAndFinish() {
         viewModelScope.launch {
             val error = AppInitializationError("No Web browser available to launch AuthFlow")
-            analyticsTracker.track(Error(Pane.UNEXPECTED_ERROR, error))
+            analyticsTracker.logError(
+                "error Launching the Auth Flow",
+                logger = logger,
+                pane = Pane.UNEXPECTED_ERROR,
+                error = error
+            )
             finishWithResult(
                 state = awaitState(),
                 result = Failed(error)
