+++ financial-connections/src/main/java/com/stripe/android/financialconnections/analytics/FinancialConnectionsEventListener.kt
@@ -1,0 +1,3 @@
+package com.stripe.android.financialconnections.analytics
+
+typealias FinancialConnectionsEventListener = (FinancialConnectionsEvent) -> Unit
