+++ financial-connections/src/main/java/com/stripe/android/financialconnections/features/institutionpicker/InstitutionPickerViewModel.kt
@@ -11,12 +11,11 @@ import com.airbnb.mvrx.ViewModelContext
 import com.stripe.android.core.Logger
 import com.stripe.android.financialconnections.FinancialConnectionsSheet
 import com.stripe.android.financialconnections.analytics.FinancialConnectionsAnalyticsTracker
-import com.stripe.android.financialconnections.analytics.FinancialConnectionsEvent
-import com.stripe.android.financialconnections.analytics.FinancialConnectionsEvent.FeaturedInstitutionsLoaded
-import com.stripe.android.financialconnections.analytics.FinancialConnectionsEvent.InstitutionSelected
-import com.stripe.android.financialconnections.analytics.FinancialConnectionsEvent.PaneLoaded
-import com.stripe.android.financialconnections.analytics.FinancialConnectionsEvent.SearchScroll
-import com.stripe.android.financialconnections.analytics.FinancialConnectionsEvent.SearchSucceeded
+import com.stripe.android.financialconnections.analytics.FinancialConnectionsInternalEvent.FeaturedInstitutionsLoaded
+import com.stripe.android.financialconnections.analytics.FinancialConnectionsInternalEvent.InstitutionSelected
+import com.stripe.android.financialconnections.analytics.FinancialConnectionsInternalEvent.PaneLoaded
+import com.stripe.android.financialconnections.analytics.FinancialConnectionsInternalEvent.SearchScroll
+import com.stripe.android.financialconnections.analytics.FinancialConnectionsInternalEvent.SearchSucceeded
 import com.stripe.android.financialconnections.analytics.logError
 import com.stripe.android.financialconnections.domain.FeaturedInstitutions
 import com.stripe.android.financialconnections.domain.GetManifest
@@ -64,8 +63,12 @@ internal class InstitutionPickerViewModel @Inject constructor(
                         ).data
                     }
                 }.onFailure {
-                    logger.error("Error fetching featured institutions", it)
-                    eventTracker.track(FinancialConnectionsEvent.Error(Pane.INSTITUTION_PICKER, it))
+                    eventTracker.logError(
+                        extraMessage = "Error fetching featured institutions",
+                        error = it,
+                        pane = Pane.INSTITUTION_PICKER,
+                        logger = logger
+                    )
                 }
 
             val (institutions, duration) = result.getOrNull() ?: Pair(emptyList(), 0L)
