+++ financial-connections/src/main/java/com/stripe/android/financialconnections/presentation/FinancialConnectionsSheetNativeViewModel.kt
@@ -15,11 +15,11 @@ import com.stripe.android.core.Logger
 import com.stripe.android.financialconnections.FinancialConnectionsSheet
 import com.stripe.android.financialconnections.R
 import com.stripe.android.financialconnections.analytics.FinancialConnectionsAnalyticsTracker
-import com.stripe.android.financialconnections.analytics.FinancialConnectionsEvent.AppBackgrounded
-import com.stripe.android.financialconnections.analytics.FinancialConnectionsEvent.ClickNavBarBack
-import com.stripe.android.financialconnections.analytics.FinancialConnectionsEvent.ClickNavBarClose
-import com.stripe.android.financialconnections.analytics.FinancialConnectionsEvent.Complete
-import com.stripe.android.financialconnections.analytics.FinancialConnectionsEvent.PaneLaunched
+import com.stripe.android.financialconnections.analytics.FinancialConnectionsInternalEvent.AppBackgrounded
+import com.stripe.android.financialconnections.analytics.FinancialConnectionsInternalEvent.ClickNavBarBack
+import com.stripe.android.financialconnections.analytics.FinancialConnectionsInternalEvent.ClickNavBarClose
+import com.stripe.android.financialconnections.analytics.FinancialConnectionsInternalEvent.Complete
+import com.stripe.android.financialconnections.analytics.FinancialConnectionsInternalEvent.PaneLaunched
 import com.stripe.android.financialconnections.di.APPLICATION_ID
 import com.stripe.android.financialconnections.di.DaggerFinancialConnectionsSheetNativeComponent
 import com.stripe.android.financialconnections.di.FinancialConnectionsSheetNativeComponent
