+++ financial-connections/src/test/java/com/stripe/android/financialconnections/FinancialConnectionsSheetViewModelTest.kt
@@ -117,7 +117,8 @@ class FinancialConnectionsSheetViewModelTest {
                     "pane" to "unexpected_error",
                     "error" to "AppInitializationError",
                     "error_type" to "AppInitializationError",
-                    "error_message" to "No Web browser available to launch AuthFlow",
+                    "error_message" to "No Web browser available to launch AuthFlow " +
+                        "error Launching the Auth Flow",
                 )
             )
         }
